Now works for 'category' taxonomy, not just 'post_tag'.

+++ primary-term.php
@@ -7,22 +7,25 @@
 class AC_Primary_Tag
 {
 	// post meta key we will use to store the primary tag id
-	const POST_META_KEY = '_ac_primary_tag';
+	const POST_META_KEY      = '_ac_primary_tag';
+	const TAXONOMY           = 'post_tag'; // tested with 'category' and 'post_tag'
+	const TAXONOMY_NICE_NAME = 'Tag';
 	
 	// Hook into actions and filters
 	static public function init() {
 		add_action( 'add_meta_boxes', array( __CLASS__, 'add_meta_box' ) );
 		add_action( 'save_post', array( __CLASS__, 'save' ) );
 		add_action( 'ac_primary_tag', array( __CLASS__, 'display_tag' ), 10, 1 );
-		add_action( 'delete_tag', array( __CLASS__, 'delete_tag' ), 10, 3 );
+		add_action( 'delete_term', array( __CLASS__, 'delete_term' ), 10, 3 );
 		add_filter( 'get_the_terms', array( __CLASS__, 'show_primary_tag_first' ), 10, 3 );
+		add_filter( 'wp_dropdown_cats', array( __CLASS__, 'append_0th_option' ), 10, 2 );
 	}
 	
 	static public function add_meta_box( $post_type ) {
 		if ( 'post' == $post_type ) {
 			add_meta_box(
 				'ac_primary_tag_meta_box' ,
-				_( 'Primary Tag' ),
+				_( 'Primary ' . self::TAXONOMY_NICE_NAME ),
 				array( __CLASS__, 'render_meta_box_content' ),
 				$post_type,
 				'advanced',
@@ -56,14 +59,8 @@ class AC_Primary_Tag
 		}
 
 		// Check the user's permissions.
-		if ( 'page' == $_POST['post_type'] ) {
-			if ( ! current_user_can( 'edit_page', $post_id ) ) {
-				return $post_id;
-			}
-		} else {
-			if ( ! current_user_can( 'edit_post', $post_id ) ) {
-				return $post_id;
-			}
+		if ( ! current_user_can( 'edit_post', $post_id ) ) {
+			return $post_id;
 		}
 
 		/* OK, its safe for us to save the data now. */
@@ -85,18 +82,6 @@ class AC_Primary_Tag
 		// Add an nonce field so we can check for it later.
 		wp_nonce_field( 'ac_primary_tag_inner_custom_box', 'ac_primary_tag_inner_custom_box_nonce' );
 
-		$tags = get_tags( array( 'hide_empty' => false ) );
-		if ( empty( $tags ) ) {
-			return;
-		}
-		
-		$tags_a = array();
-		foreach ( $tags as $tag ) {
-			$tags_a[ $tag->term_id ] = $tag->name;
-		}
-		$none_selected = array( 0 => 'Select a Primary Tag' );
-		$tags_a = $none_selected + $tags_a;
-
 		// Use get_post_meta to retrieve an existing value from the database.
 		$current_value = self::get_primary_tag( $post->ID );
 		
@@ -113,14 +98,15 @@ class AC_Primary_Tag
 		
 		// Display the form, using the current value.
 		echo '<label for="ac_primary_tag_field">';
-		_e( 'Primary Tag' );
+		_e( 'Primary ' . self::TAXONOMY_NICE_NAME );
 		echo '</label> ';
-		echo '<select type="text" id="ac_primary_tag_field" name="ac_primary_tag_field">';
-		foreach ( $tags_a as $term_id => $term_name ) {
-			$selected = $term_id == $current_value ? 'selected="selected"' : '';
-			echo sprintf( '<option value="%d" %s />%s</option>', intval( $term_id ), intval( $selected ), esc_attr( $term_name ) );
-		}
-		echo '</select>';
+		wp_dropdown_categories( array(
+			'taxonomy'     => self::TAXONOMY,
+			'hide_empty'   => false,
+			'hierarchical' => true,
+			'class'        => 'ac_primary_term_dropdown',
+			'selected'     => $current_value,
+		) );
 	}
 	
 	static public function display_tag( $show_as_link = false ) {
@@ -136,12 +122,20 @@ class AC_Primary_Tag
 		
 		$primary_tag_id = self::get_primary_tag( $post->ID );
 		
-		$primary_tag = get_term( $primary_tag_id, 'post_tag' );
+		if ( empty( $primary_tag_id ) ) {
+			return;
+		}
 		
-		if ( false ==  $show_as_link ) {
+		$primary_tag = get_term( $primary_tag_id, self::TAXONOMY );
+		
+		if ( empty( $primary_tag ) ) {
+			return;
+		}
+		
+		if ( false == $show_as_link ) {
 			return $primary_tag->name;
 		} else {
-			$permalink = get_term_link( $primary_tag, 'post_tag' );
+			$permalink = get_term_link( $primary_tag, self::TAXONOMY );
 			return sprintf( '<a href="%s">%s</a>', $permalink, $primary_tag->name );
 		}
 	}
@@ -154,21 +148,25 @@ class AC_Primary_Tag
 	 * @param mixed   $deleted_term Copy of the already-deleted term, in the form specified
 	 *                              by the parent function. WP_Error otherwise.
 	 */
-	static public function delete_tag( $term, $tt_id, $deleted_term ) {
+	static public function delete_term( $term, $tt_id, $taxonomy, $deleted_term ) {
+		if ( self::TAXONOMY != $taxonomy ) {
+			return;
+		}
+
 		global $wpdb;
 		// WordPress already verified that the user has permissions to delete the term so we'll execute the query
 		$wpdb->query( $wpdb->prepare( "DELETE FROM $wpdb->postmeta WHERE meta_key='%s' AND meta_value=%d", self::POST_META_KEY, $term ) );
 	}
 	
 	static public function get_term( $term_id ) {
-		$term = get_term_by( 'id',  (int) $term_id, 'post_tag' );
+		$term = get_term_by( 'id',  (int) $term_id, self::TAXONOMY );
 		if ( $term ) {
 			return $term;
 		} else {
 			return false;
 		}
 	}
-	
+
 	/**
 	* Filter the list of terms attached to the given post.
 	*
@@ -177,7 +175,7 @@ class AC_Primary_Tag
 	* @param string $taxonomy Name of the taxonomy.
 	*/
 	static public function show_primary_tag_first( $terms, $post_id, $taxonomy ) {
-		if ( 'post_tag' != $taxonomy ) {
+		if ( self::TAXONOMY != $taxonomy ) {
 			return $terms;
 		}
 		$current_value = self::get_primary_tag( $post_id );
@@ -206,7 +204,51 @@ class AC_Primary_Tag
 		// merge 
 		return array_merge( $primary_term, $all_other_terms );
 	}
-	
+
+	static public	function append_0th_option( $output, $r ) {
+		if ( 'ac_primary_term_dropdown' != $r['class'] ) {
+			return $output;
+		}
+
+		$doc = new DOMDocument();
+		$output = str_replace( '&nbsp;', '-', $output );
+		$doc->loadXML( $output );
+
+		$parent_path = '//select';
+		$next_path = '//select/*[1]';
+
+		// Find the parent node
+		$xpath = new DomXPath( $doc );
+
+		// Find parent node
+		$parent = $xpath->query( $parent_path );
+
+		// Append name and id to select
+		$select = $doc->getElementById( 'cat' );
+		$select = $xpath->query( "//*[@id='cat']" )->item( 0 );
+		$select->setAttribute( 'name', 'ac_primary_tag_field' );
+		$select->setAttribute( 'id',   'ac_primary_tag_field' );
+
+		// new node will be inserted before this node
+		$next = $xpath->query( $next_path );
+
+		// Create the new element
+		$element = $doc->createElement( 'option', 'Please select a default term' );
+
+		$domAttribute = $doc->createAttribute( 'value' );
+
+		// Value for the created attribute
+		$domAttribute->value = '0';
+
+		// Don't forget to append it to the element
+		$element->appendChild( $domAttribute );
+
+		// Insert the new element
+		$parent->item( 0 )->insertBefore( $element, $next->item( 0 ) );
+
+		return $doc->saveXML();
+	}
+
 	/**
 	 * Get primary tag of a post
 	 *
